+++ docs/src/conf.py
@@ -57,7 +57,7 @@ intersphinx_mapping = {
     "python": ("https://docs.python.org/3", None),
     "torch": ("https://pytorch.org/docs/stable/", None),
     "e3nn": ("https://docs.e3nn.org/en/latest/", None),
-    "metatensor": ("https://docs.metatensor.org/latest/index.html", None),
+    "metatensor": ("https://docs.metatensor.org/latest/", None),
 }
 
 html_theme = "furo"
