+++ python/src/sphericart/__init__.py
@@ -1,2 +1,1 @@
 from .spherical_harmonics import SphericalHarmonics, SolidHarmonics  # noqa
-from . import metatensor  # noqa
