+++ python/src/sphericart/metatensor.py
@@ -1,20 +1,17 @@
-import torch
+from typing import List, Optional
 
-from .. import SolidHarmonics as RawSolidHarmonics
-from .. import SphericalHarmonics as RawSphericalHarmonics
-from sphericart.metatensor.spherical_harmonics import (
-    _check_xyz_tensor_map,
-    _wrap_into_tensor_map,
-)
+import numpy as np
+
+from .spherical_harmonics import SolidHarmonics as RawSolidHarmonics
+from .spherical_harmonics import SphericalHarmonics as RawSphericalHarmonics
 
 
 try:
-    import metatensor.torch
-    from metatensor.torch import Labels, TensorMap
+    import metatensor
+    from metatensor import Labels, TensorMap
 except ImportError as e:
     raise ImportError(
-        "the `sphericart.torch.metatensor` module requires "
-        "`metatensor-torch` to be installed"
+        "the `sphericart.metatensor` module requires `metatensor` to be installed"
     ) from e
 
 
@@ -27,22 +24,22 @@ class SphericalHarmonics:
         # precompute some labels
         self.precomputed_keys = Labels(
             names=["o3_lambda"],
-            values=torch.arange(l_max + 1).reshape(-1, 1),
+            values=np.arange(l_max + 1).reshape(-1, 1),
         )
         self.precomputed_mu_components = [
             Labels(
                 names=["o3_mu"],
-                values=torch.arange(-l, l + 1).reshape(-1, 1),
+                values=np.arange(-l, l + 1).reshape(-1, 1),
             )
             for l in range(l_max + 1)  # noqa E741
         ]
         self.precomputed_xyz_components = Labels(
             names=["xyz"],
-            values=torch.arange(3).reshape(-1, 1),
+            values=np.arange(3).reshape(-1, 1),
         )
         self.precomputed_xyz_2_components = Labels(
             names=["xyz_2"],
-            values=torch.arange(3).reshape(-1, 1),
+            values=np.arange(3).reshape(-1, 1),
         )
         self.precomputed_properties = Labels.single()
 
@@ -57,7 +54,6 @@ class SphericalHarmonics:
             self.precomputed_xyz_components,
             self.precomputed_xyz_2_components,
             self.precomputed_properties,
-            metatensor_module=metatensor.torch,
         )
 
     def compute_with_gradients(self, xyz: TensorMap) -> TensorMap:
@@ -74,7 +70,6 @@ class SphericalHarmonics:
             self.precomputed_xyz_2_components,
             self.precomputed_properties,
             sh_gradients,
-            metatensor_module=metatensor.torch,
         )
 
     def compute_with_hessians(self, xyz: TensorMap) -> TensorMap:
@@ -87,12 +82,11 @@ class SphericalHarmonics:
             self.precomputed_keys,
             xyz.block().samples,
             self.precomputed_mu_components,
-            self.precomputed_properties,
             self.precomputed_xyz_components,
             self.precomputed_xyz_2_components,
+            self.precomputed_properties,
             sh_gradients,
             sh_hessians,
-            metatensor_module=metatensor.torch,
         )
 
 
@@ -105,26 +99,26 @@ class SolidHarmonics:
         # precompute some labels
         self.precomputed_keys = Labels(
             names=["o3_lambda"],
-            values=torch.arange(l_max + 1).reshape(-1, 1),
+            values=np.arange(l_max + 1).reshape(-1, 1),
         )
         self.precomputed_mu_components = [
             Labels(
                 names=["o3_mu"],
-                values=torch.arange(-l, l + 1).reshape(-1, 1),
+                values=np.arange(-l, l + 1).reshape(-1, 1),
             )
             for l in range(l_max + 1)  # noqa E741
         ]
         self.precomputed_xyz_components = Labels(
             names=["xyz"],
-            values=torch.arange(3).reshape(-1, 1),
+            values=np.arange(3).reshape(-1, 1),
         )
         self.precomputed_xyz_2_components = Labels(
             names=["xyz_2"],
-            values=torch.arange(3).reshape(-1, 1),
+            values=np.arange(3).reshape(-1, 1),
         )
         self.precomputed_properties = Labels.single()
 
-    def compute(self, xyz: torch.Tensor) -> TensorMap:
+    def compute(self, xyz: np.ndarray) -> TensorMap:
         _check_xyz_tensor_map(xyz)
         sh_values = self.raw_calculator.compute(xyz.block().values.squeeze(-1))
         return _wrap_into_tensor_map(
@@ -135,10 +129,9 @@ class SolidHarmonics:
             self.precomputed_xyz_components,
             self.precomputed_xyz_2_components,
             self.precomputed_properties,
-            metatensor_module=metatensor.torch,
         )
 
-    def compute_with_gradients(self, xyz: torch.Tensor) -> TensorMap:
+    def compute_with_gradients(self, xyz: np.ndarray) -> TensorMap:
         _check_xyz_tensor_map(xyz)
         sh_values, sh_gradients = self.raw_calculator.compute_with_gradients(
             xyz.block().values.squeeze(-1)
@@ -152,10 +145,9 @@ class SolidHarmonics:
             self.precomputed_xyz_2_components,
             self.precomputed_properties,
             sh_gradients,
-            metatensor_module=metatensor.torch,
         )
 
-    def compute_with_hessians(self, xyz: torch.Tensor) -> TensorMap:
+    def compute_with_hessians(self, xyz: np.ndarray) -> TensorMap:
         _check_xyz_tensor_map(xyz)
         sh_values, sh_gradients, sh_hessians = (
             self.raw_calculator.compute_with_hessians(xyz.block().values.squeeze(-1))
@@ -170,5 +162,68 @@ class SolidHarmonics:
             self.precomputed_properties,
             sh_gradients,
             sh_hessians,
-            metatensor_module=metatensor.torch,
         )
+
+
+def _check_xyz_tensor_map(xyz: TensorMap):
+    if len(xyz.blocks()) != 1:
+        raise ValueError("`xyz` should have only one block")
+    if len(xyz.block().components) != 1:
+        raise ValueError("`xyz` should have only one component")
+    if xyz.block().components[0].names != ["xyz"]:
+        raise ValueError("`xyz` should have only one component named 'xyz'")
+    if xyz.block().components[0].values.shape[0] != 3:
+        raise ValueError("`xyz` should have 3 Cartesian coordinates")
+    if xyz.block().properties.values.shape[0] != 1:
+        raise ValueError("`xyz` should have only one property")
+
+
+def _wrap_into_tensor_map(
+    sh_values: np.ndarray,
+    keys: Labels,
+    samples: Labels,
+    components: List[Labels],
+    xyz_components: Labels,
+    xyz_2_components: Labels,
+    properties: Labels,
+    sh_gradients: Optional[np.ndarray] = None,
+    sh_hessians: Optional[np.ndarray] = None,
+) -> TensorMap:
+
+    # infer l_max
+    l_max = len(components) - 1
+
+    blocks = []
+    for l in range(l_max + 1):  # noqa E741
+        l_start = l**2
+        l_end = (l + 1) ** 2
+        sh_values_block = metatensor.TensorBlock(
+            values=sh_values[:, l_start:l_end, None],
+            samples=samples,
+            components=[components[l]],
+            properties=properties,
+        )
+        if sh_gradients is not None:
+            sh_gradients_block = metatensor.TensorBlock(
+                values=sh_gradients[:, :, l_start:l_end, None],
+                samples=samples,
+                components=[xyz_components, components[l]],
+                properties=properties,
+            )
+            if sh_hessians is not None:
+                sh_hessians_block = metatensor.TensorBlock(
+                    values=sh_hessians[:, :, :, l_start:l_end, None],
+                    samples=samples,
+                    components=[
+                        xyz_2_components,
+                        xyz_components,
+                        components[l],
+                    ],
+                    properties=properties,
+                )
+                sh_gradients_block.add_gradient("positions", sh_hessians_block)
+            sh_values_block.add_gradient("positions", sh_gradients_block)
+
+        blocks.append(sh_values_block)
+
+    return metatensor.TensorMap(keys=keys, blocks=blocks)
