--- python/src/sphericart/metatensor/__init__.py
@@ -1,1 +1,0 @@
-from .spherical_harmonics import SphericalHarmonics, SolidHarmonics  # noqa
